+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ## Unreleased
 
 * Update PowerSTIG to to parse/convert the Windows Server 2012 R2 DC V2R19: [#578](https://github.com/microsoft/PowerStig/issues/578)
+* Update PowerSTIG to parse/convert the Windows Defender STIG V1R7: [576](https://github.com/microsoft/PowerStig/issues/576)
+* Update PowerSTIG to successfully parse Mozilla Firefox STIG - Ver 4, Rel 28: [#573](https://github.com/microsoft/PowerStig/issues/573)
+* Update PowerSTIG to parse and apply Adobe Acrobat Reader Version 1, Release 6: [562](https://github.com/microsoft/PowerStig/issues/562)
 * Update PowerSTIG release process to include STIG Coverage markdown wiki automation: [#560](https://github.com/microsoft/PowerStig/issues/560)
 * Fixed [#259](https://github.com/microsoft/PowerStig/issues/259): Checklist .ckl file fails XML validation in Stig Viewer 2.8.
 * Fixed [#527](https://github.com/microsoft/PowerStig/issues/527): Checklist is not using manualcheckfile when using DscResult.